Bug fix, add "web page title" to web pages for future ranking
Bug fix: re-add mutex for _totalPagesScraped int. Without the mutex lock we have threading issues which end in way more pages being added to the repository than intended

Add logic to scrape and save web page titles while we process and index pages. webPageTitle was added to the web page class, added method to web page repository to get the web page title, refactored search engine search method to account for the web page title, refactor search engine main to store top search results in a tuple instead of pair to account for new variable, add logic to web page processor to actually pull the web page title from the DOM tree. The end goal is to use the web page title to assist in ranking by determining the context of web pages based on the title AND content, not just the number of times certain words appear on a page