*BUG* but also add more filtering to url scraping
Modify method that scrapes web page URLs to remove query parameters that would slightly modify the URL so I'd index the same url tons of times, there is a bug that is likely due to multi threading at the moment. If I try to scrape 1000 pages the program thinks it scraped 1000 pages but the _index only shows 900-980 web pages. I plan to implement a thread pool to reuse threads and hopefully utilize more locks inside the threadpool to prevent data racing